+++ src/datasources/asset/data-sources/list-assets/ListAssetsDataSource.ts
@@ -1,10 +1,10 @@
 import { DataQueryRequest, DataFrameDTO, DataSourceInstanceSettings } from '@grafana/data';
-import { AssetDataSourceOptions, AssetQuery, AssetQueryType } from '../../../types/types';
-import { AssetModel, AssetsResponse } from '../../../../asset-common/types';
-import { getWorkspaceName } from '../../../../../core/utils';
 import { BackendSrv, getBackendSrv, getTemplateSrv, TemplateSrv } from '@grafana/runtime';
 import { AssetDataSourceBase } from '../AssetDataSourceBase';
-import { ListAssetsQuery } from '../../../types/ListAssets.types';
+import { AssetDataSourceOptions, AssetQuery, AssetQueryType } from '../../types/types';
+import { ListAssetsQuery } from '../../types/ListAssets.types';
+import { AssetModel, AssetsResponse } from '../../../asset-common/types';
+import { getWorkspaceName } from '../../../../core/utils';
 
 export class ListAssetsDataSource extends AssetDataSourceBase {
   private dependenciesLoadedPromise: Promise<void>;