+++ src/datasources/asset/defaults.ts
@@ -3,6 +3,8 @@ export const defaultAssetSummaryQuery = {
 }
 
 export const defaultCalibrationForecastQuery = {
+    filter: "",
+    groupBy: []
 }
 
 export const defaultListAssetsQuery = {